Support "exotic" arguments in C++ func registration.  These are in-place returns, async handles and FP12 structs.  This also makes the support for range arguments much more natural.  Some gymnastics in ExcelRef.h to allow RangeArg to have range-like funcationality.
tests: add exotic argument functions to TestAddin

+++ libs/xlOil_Utils/xloSplit.cpp
@@ -86,8 +86,8 @@ namespace xloil
       bool byRow = inputArray.nCols() == 1;
 
       ExcelArrayBuilder builder(
-        byRow ? inputArray.size() : maxTokens, 
-        byRow ? maxTokens : inputArray.size(),
+        byRow ? inputArray.size() : (int)maxTokens, 
+        byRow ? (int)maxTokens : inputArray.size(),
         totalStrLength);
 
       // We don't intend to write to every cell, so need to initialise
